+++ outlook3/src/main/resources/docs/pages/release-notes.md
@@ -1,6 +1,6 @@
 # Release Notes – Krista Outlook Extension
 
-## Version 3.0.7
+## Version 3.0.8
 
 - **Developers**: VaraPrasad Kolli, Vaibhav Choudhary
 - **Krista Service APIs (Java)**: 1.0.115
@@ -10,6 +10,15 @@
 ---
 
 ## Resolved Bugs
+
+- [**KE-2358**](https://antbrains.atlassian.net/browse/KE-2358):  
+  Outlook extension Paging occasionally fails on when iterating over pages - Fetch Mails By Label request
+
+- [**KE-2349**](https://antbrains.atlassian.net/browse/KE-2349):  
+  Outlook Extension: Fetch Latest Mail request should fetch the latest mail object
+
+- [**KE-2273**](https://antbrains.atlassian.net/browse/KE-2273):  
+  Outlook Extension, Add "include email thread" option in the Reply to Mail and Forward Mail requests
 
 - [**KE-2293**](https://antbrains.atlassian.net/browse/KE-2293):  
   Our outlook extension has limits on the max Page Number and max Page Size parameters for the "Fetch Mails By Label"